+++ src/app/app.component.html
@@ -1,122 +1,12 @@
 v24021611
-<br>
-<br>
-<div class="demo-button-container ">
-  <mat-form-field appearance="outline" class="demo-button">
-    <mat-label>Name</mat-label>
-    <input  [formControl]="name" matInput placeholder="">
-    <mat-icon matSuffix>sentiment_very_satisfied</mat-icon>
-  </mat-form-field>
-  <mat-form-field appearance="outline" class="demo-button amount-form-field">
-    <mat-label>Amount</mat-label>
-    <input [formControl]="amount"
-    currencyMask
-    [options]="{ prefix: 'R$ ', thousands: '.', decimal: ',' }"
-      matInput 
-    type="text" 
-    class="example-right-align"
-    placeholder="0">
-  </mat-form-field>
-  <mat-slide-toggle [formControl]="free" class="demo-button">Free</mat-slide-toggle>
-  <br>
-  <button mat-raised-button (click)="addData()" class="demo-button">
-    Add data
-  </button>
-  <button mat-raised-button (click)="clearFields()">Clear fields</button>
-  <button
-  mat-raised-button
-  [disabled]="!dataToDisplay.length"
-  (click)="removeData()"
-  class="demo-button">
-  Remove select data
-</button>
-<button mat-raised-button (click)="clearTable()">Clear table</button>
-<button mat-raised-button (click)="updateCurrentBalance()">Update Current Balance</button>
-<button mat-raised-button (click)="simulation()">Simulation</button>
+<mat-tab-group>
+  <mat-tab label="Refund"> refund </mat-tab>
+  <mat-tab label="Balance">
+    <br>
+    <br>
+    <app-balance-component></app-balance-component>
+  </mat-tab>
 
-</div>
-<table mat-table [dataSource]="dataSource" class="mat-elevation-z8" matSort (matSortChange)="announceSortChange($event)">
-  
-  <!-- Checkbox Column -->
-  <ng-container matColumnDef="select">
-    <th mat-header-cell *matHeaderCellDef>
-      <mat-checkbox (change)="$event ? toggleAllRows() : null"
-                    [checked]="selection.hasValue() && isAllSelected()"
-                    [indeterminate]="selection.hasValue() && !isAllSelected()"
-                    [aria-label]="checkboxLabel()">
-      </mat-checkbox>
-    </th>
-    <td mat-cell *matCellDef="let row">
-      <mat-checkbox (click)="$event.stopPropagation()"
-                    (change)="$event ? selection.toggle(row) : null"
-                    [checked]="selection.isSelected(row)"
-                    [aria-label]="checkboxLabel(row)">
-      </mat-checkbox>
-    </td>
-    <td mat-footer-cell *matFooterCellDef> Total </td>
-  </ng-container>
+</mat-tab-group>
 
-  <!-- Idplayer Column -->
-  <ng-container matColumnDef="idplayer">
-    <th mat-header-cell *matHeaderCellDef mat-sort-header sortActionDescription="Sort by idplayer"> No. </th>
-    <td mat-cell *matCellDef="let element"> {{element.idplayer}} </td>
-    <td mat-footer-cell *matFooterCellDef> {{total.value}} </td>
-  </ng-container>
 
-  <!-- Name Column -->
-  <ng-container matColumnDef="name">
-    <th mat-header-cell *matHeaderCellDef mat-sort-header sortActionDescription="Sort by name"> Name </th>
-    <td mat-cell *matCellDef="let element"> {{element.name}} </td>
-    <td mat-footer-cell *matFooterCellDef>{{totalForPerson.value | currency}} for person</td>
-  </ng-container>
-
-  <!-- Amount Column -->
-  <ng-container matColumnDef="amount">
-    <th mat-header-cell *matHeaderCellDef mat-sort-header sortActionDescription="Sort by amount"> Amount </th>
-    <td mat-cell *matCellDef="let element"> {{element.amount | currency}} </td>
-    <td mat-footer-cell *matFooterCellDef>{{totalAmount.value | currency}}</td>
-  </ng-container>
-
-  <!-- Balance Column -->
-  <ng-container matColumnDef="balance">
-    <th mat-header-cell *matHeaderCellDef mat-sort-header sortActionDescription="Sort by balance"> Balance </th>
-    <td mat-cell *matCellDef="let element"> {{element.balance | currency}} </td>
-    <td mat-footer-cell *matFooterCellDef>{{totalBalance.value | currency}}</td>
-  </ng-container>
-
-  <!-- Positive Current Balance Column -->
-  <ng-container matColumnDef="positive_current_balance">
-    <th mat-header-cell *matHeaderCellDef mat-sort-header sortActionDescription="Sort by positive_current_balance"> Players </th>
-    <td mat-cell *matCellDef="let element">
-      @for (color of element.players; track player) {
-        <div>{{ color.name }} {{ color.balance  | currency}}</div>
-        } </td>
-    <td mat-footer-cell *matFooterCellDef></td>
-  </ng-container>
-
-  <!-- Current Balance Column -->
-  <ng-container matColumnDef="current_balance">
-    <th mat-header-cell *matHeaderCellDef mat-sort-header sortActionDescription="Sort by current_balance"> Current Balance </th>
-    <td mat-cell *matCellDef="let element"> {{element.current_balance | currency}} </td>
-    <td mat-footer-cell *matFooterCellDef>{{totalCurrentBalance.value | currency}}</td>
-  </ng-container>
-
-  <!-- Free Column -->
-  <ng-container matColumnDef="free">
-    <th mat-header-cell *matHeaderCellDef mat-sort-header sortActionDescription="Sort by free"> Free </th>
-    <td mat-cell *matCellDef="let element"> 
-      @if (element.free) {
-        <mat-icon aria-hidden="false" aria-label="icon" fontIcon="done"></mat-icon>
-      } @else {
-        <mat-icon aria-hidden="false" aria-label="icon" fontIcon="close"></mat-icon>
-      }
-     </td>
-     <td mat-footer-cell *matFooterCellDef>{{totalNoFree.value}} no free</td>
-    </ng-container>
-
-  <tr mat-header-row *matHeaderRowDef="displayedColumns"></tr>
-  <tr mat-row *matRowDef="let row; columns: displayedColumns;"
-      (click)="selection.toggle(row)">
-  </tr>
-  <tr mat-footer-row *matFooterRowDef="displayedColumns; sticky: true"></tr>
-</table>
