+++ src/app/app.module.ts
@@ -11,13 +11,16 @@ import { MatIconModule } from '@angular/material/icon';
 import { MatInputModule } from '@angular/material/input';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatSlideToggleModule } from '@angular/material/slide-toggle';
-import { MatSortModule} from '@angular/material/sort';
+import { MatTabsModule } from '@angular/material/tabs';
+import { MatSortModule } from '@angular/material/sort';
 import { ReactiveFormsModule } from '@angular/forms';
 import { NgxCurrencyDirective } from 'ngx-currency';
+import { BalanceComponentComponent } from './balance-component/balance-component.component';
 
 @NgModule({
   declarations: [
-    AppComponent
+    AppComponent,
+    BalanceComponentComponent
   ],
   imports: [
     BrowserModule,
@@ -31,6 +34,7 @@ import { NgxCurrencyDirective } from 'ngx-currency';
     MatFormFieldModule,
     MatSlideToggleModule,
     MatSortModule,
+    MatTabsModule,
     ReactiveFormsModule,
     NgxCurrencyDirective
   ],
